Normalize the integration test environment
This change (partially) normalizes the environment that integration tests
are run under. An unrestricted environment can break at least two tests
due to GIT_EDITOR being set, which overrides the core.editor config used
for test isolation. A whitelist of environment variables that are either
set by the test harness or which must be picked up from the user environment
is established, with a function that returns the filtered environment.