Make env var allowlist even smaller
The purpose of the allowlist is to determine which env vars are inherited
from the host environment. 'GIT_CONFIG_GLOBAL' should never be inherited
from the host environment because we don't want to be using the host's git
config.

The reason we had this included previously is that we were using os.Setenv
to set the env var in the test runner process, but we really should have just
passed the env var explicitly wherever it was needed (i.e. in the shell and
in the lazygit process). So this commit does that.

One small nuance: there are two places we use the shell struct: first when coming up
with the text fixture, and then inside the lazygit process. In the latter, we're
using os.Environ() because it's already filtered down by the test runner,
however that does mean that technically it's going to have some additional
env vars that aren't defined when we're creating the text fixture. But
this should not matter because the additional env vars are our own lazygit-
specific env vars